moved classes to jersey packages
