+++ solr/core/src/java/org/apache/solr/jersey/APIConfigProvider.java
@@ -1,0 +1,34 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.solr.jersey;
+
+import org.apache.solr.jersey.APIConfigProvider.APIConfig;
+import org.glassfish.hk2.api.Factory;
+
+/**
+ * Interface to be implemented by the Request Handlers that need to provide some custom
+ * configuration to the V2 APIs
+ */
+public interface APIConfigProvider<T extends APIConfig> extends Factory<T> {
+
+  @Override
+  default void dispose(T instance) {}
+
+  Class<T> getConfigClass();
+
+  public interface APIConfig {}
+}
